==> notebooks/00_00_v2_wuserade_download_poem.ipynb <==
import requests

class UseradeClient:
    def __init__(self, base_url="https://wuserade.onrender.com"):
        self.base_url = base_url

    def get_poems(self, page=1):
        """Получение списка стихов."""
        url = f"{self.base_url}/poems"
        params = {"page": page}
        response = requests.get(url, params=params)
        return response.json()

    def get_poem_by_id(self, poem_id):
        """Получение стиха по ID."""
        url = f"{self.base_url}/poems/{poem_id}"
        response = requests.get(url)
        return response.json()

    def get_authors(self, page=1):
        """Получение списка авторов."""
        url = f"{self.base_url}/authors"
        params = {"page": page}
        response = requests.get(url, params=params)
        return response.json()

    def get_author_by_id(self, author_id):
        """Получение информации об авторе по ID."""
        url = f"{self.base_url}/authors/{author_id}"
        response = requests.get(url)
        return response.json()

    def get_poems_of_author(self, author_id):
        """Получение стихов конкретного автора."""
        url = f"{self.base_url}/authors/{author_id}/poems"
        response = requests.get(url)
        return response.json()

    def search_poems(self, search_query):
        """Поиск стихов."""
        url = f"{self.base_url}/poems/search"
        params = {"q": search_query}
        response = requests.get(url, params=params)
        return response.json()

    def get_themes(self):
        """Получение списка тем."""
        url = f"{self.base_url}/themes"
        response = requests.get(url)
        return response.json()
    
    def all_poems(self):
        """Генератор для получения всех стихов, обходя пагинацию."""
        current_page = 1
        total_pages = 1  # Начинаем с 1, чтобы войти в цикл

        while current_page <= total_pages:
            response = self.get_poems(page=current_page)
            if 'total_pages' in response:
                total_pages = response['total_pages']

            if 'poems' in response:  # Предполагаем, что стихи возвращаются в ключе 'data'
                for poem in response['poems']:
                    yield poem

            current_page += 1
    
    # Здесь добавьте оставшиеся методы, если они есть.

# Пример использования
client = UseradeClient()

poems = {}

for poem in client.all_poems():
    poems[poem['id']] = poem

len(poems)

import json

with open('../data/external/wuserade/poems.json', 'w', encoding='utf-8') as f:
    json.dump(poems, f, ensure_ascii=False, indent=4)

import json
with open('../data/external/wuserade/poems.json', 'r', encoding='utf-8') as f:
    poems = json.load(f)

import pandas as pd

df = pd.DataFrame(poems).T

df['author_id'] = df['author'].apply(lambda x: x['id'])
df['author_name'] = df['author'].apply(lambda x: x['name'])

df = df.drop(columns=['author'])

df['content'] = df['content'].apply(lambda x: x.replace('ӏ', 'I'))
df['content_len'] = df['content'].apply(lambda x: len(x))

import nltk

df['sentences'] = df['content'].apply(lambda x: nltk.sent_tokenize(x))
df['sentences_count'] = df['sentences'].apply(len)

df

df

from tqdm import tqdm
from sentence_transformers import SentenceTransformer

model = SentenceTransformer('./sbert_from_mlm_bert_8')

vectors = [
    model.encode(sent)
    for sent in tqdm(df['content'].tolist())
]

from sklearn.cluster import AgglomerativeClustering
from collections import defaultdict


def get_clusters(vectors, n_clusters=20):
    agg_clustering = AgglomerativeClustering(n_clusters=n_clusters)
    labels = agg_clustering.fit_predict(vectors)
    return labels


def get_sents_by_clusters(words, labels):
    sents_by_clusters = defaultdict(list)
    for i, label in enumerate(labels):
        sents_by_clusters[label].append(words[i])

    return sents_by_clusters

n_clusters = 1000
labels = get_clusters(vectors, n_clusters=n_clusters)
sents_by_clusters = get_sents_by_clusters(df['id'], labels)
df['cluster'] = labels

for cluster_label, cluster_sents in sents_by_clusters.items():
    print(f'Cluster {cluster_label}:')
    print(cluster_sents)
    print()

import os

poems_cluster_dir = f'../data/processed/userade/poems_cluster/{n_clusters}'
os.makedirs(poems_cluster_dir, exist_ok=True)

for group_name, group in df.groupby('cluster'):
    group = group.drop(columns=['cluster'])
    group.to_csv(f'{poems_cluster_dir}/{group_name}.csv', index=False)
