==> notebooks/00_clean_ocr_text.ipynb <==
import os

from src.logic.text_cleaner import clean_text

external_dir = '../data/external/book'
interim_dir = '../data/interim/book'
os.makedirs(interim_dir, exist_ok=True)

processed_dir = '../data/processed/book'
os.makedirs(processed_dir, exist_ok=True)

for file_name in os.listdir(external_dir):
    with open(os.path.join(external_dir, file_name), 'r') as f:
        text = f.read()
    
    text = clean_text(text)
    
    with open(os.path.join(processed_dir, file_name), 'w') as f:
        f.write(text)
