==> notebooks/00_clean_pdf_text.ipynb <==
import os
from src.logic.text_cleaner import clean_text


from tqdm import tqdm

txt_dir = '../data/interim/smikbr_ru_txt'
output_dir = '../data/processed/smikbr_ru_txt'
os.makedirs(output_dir, exist_ok=True)

for file in tqdm(os.listdir(txt_dir)):
    if file.endswith(".txt"):
        with open(f"{txt_dir}/{file}", "r") as f:
            text = f.read()
        
        text = clean_text(text)
        with open(f"{output_dir}/{file}", "w") as f:
            f.write(text)

groups = [
    'goryanka',
    'oshamaho',
    'mingit',
    'litkb',
    'kbp',
    'zaman',
    'adigps',
    'sm'
]

for group in groups:
    os.makedirs(os.path.join(output_dir, group), exist_ok=True)

for file in tqdm(os.listdir(output_dir)):
    if os.path.isdir(file):
        continue

    for group in groups:
        if group in file.replace('smikbr.ru_arhiv', ''):
            try:
                os.rename(f"{output_dir}/{file}", f"{output_dir}/{group}/{file}")
                break
            except:
                pass
