==> notebooks/00_clean_scraped_text.ipynb <==
import os

import pandas as pd

from src.logic.text_cleaner import clean_text


external_dir = '../data/external/'
interim_dir = '../data/interim/'
os.makedirs(interim_dir, exist_ok=True)

processed_dir = '../data/processed/'
os.makedirs(processed_dir, exist_ok=True)

for file_name in ['apkbr_ru.jsonl', 'elgkbr_ru.jsonl']:
    df = pd.read_json(os.path.join(external_dir, file_name), lines=True)
    df['content'] = df['content'].apply(clean_text)

    df.to_json(
        os.path.join(interim_dir, file_name),
        orient='records',
        lines=True,
        force_ascii=False
    )

    text = '\n'.join(df['content'].tolist())
    with open(os.path.join(processed_dir, file_name.replace('.jsonl', '.txt')), 'w') as f:
        f.write(text)
