==> notebooks/01_extract_text_from_pdf.ipynb <==
import os

import textract
from tqdm import tqdm

from src.logic.text_cleaner import clean_text

def extract_text_from_pdf_file(pdf_path):
    try:
        text = textract.process(pdf_path, method='pdfminer')
    except Exception:
        return ''

    text = text.decode('utf-8')
    return text

oshhamaho_external_pdf_dir = '../data/external/pdf'
oshhamaho_interim_txt_dir = '../data/interim/pdf'
os.makedirs(oshhamaho_interim_txt_dir, exist_ok=True)

pdf_files = [f for f in os.listdir(oshhamaho_external_pdf_dir) if f.endswith('.pdf')]

pgbar = tqdm(pdf_files)

for pdf_file in pgbar:
    pgbar.set_description(pdf_file)
    pdf_path = os.path.join(oshhamaho_external_pdf_dir, pdf_file)

    text_i = extract_text_from_pdf_file(pdf_path)
    cleaned_text = clean_text(text_i)

    text_file = pdf_file.replace('.pdf', '.txt')
    text_path = os.path.join(oshhamaho_interim_txt_dir, text_file)
    with open(text_path, 'w') as f:
        f.write(cleaned_text)

txt_files = sorted([f for f in os.listdir(oshhamaho_interim_txt_dir) if f.endswith('.txt')])
processed_dir = '../data/processed/'
os.makedirs(processed_dir, exist_ok=True)

oshhamaho_all_txt_path = os.path.join(processed_dir, 'oshhamaho.txt')

pgbar = tqdm(txt_files)
text = ''

for txt_file in pgbar:
    pgbar.set_description(txt_file)
    txt_path = os.path.join(oshhamaho_interim_txt_dir, txt_file)

    with open(txt_path, 'r') as f:
        text_i = f.read()

    cleaned_text = clean_text(text_i)
    text += cleaned_text + '\n'

with open(oshhamaho_all_txt_path, 'w') as f:
    f.write(text)
