==> notebooks/02_00_select_sent_by_word.ipynb <==
import nltk
from tqdm import tqdm
from collections import defaultdict
import pandas as pd

word_df = pd.read_csv('../data/processed/word_freqs/freq_1000000_oshhamaho.csv')
word_df = word_df.head(30000)

with open('../data/processed/oshhamaho.txt', 'r') as f:
    text = f.read()

short_sents = set([sent for sent in nltk.sent_tokenize(text) if 60 < len(sent) < 80]) 
short_sent_uniq = set()

word_sent = defaultdict(list)
for word in tqdm(word_df['word']):
    for sent in short_sents:
        if word in sent:
            word_sent[word].append(sent)
            short_sent_uniq.add(sent)
        if len(word_sent[word]) > 10:
            break

with open('../data/processed/short_sents.txt', 'w') as f:
    f.write('\n'.join(sorted(short_sent_uniq)))

word_sent_df = pd.DataFrame.from_dict(word_sent, orient='index').T
word_sent_df = word_sent_df.head(10)
word_sent_df = word_sent_df.T

word_sent_df
