==> notebooks/02_00_select_sent_from_glossary.ipynb <==
import pandas as pd

rus_kbd_glossary_df = pd.read_csv('../data/processed/glossary/html_cleaned/Rus-Ady_UASP.csv')
kbd_rus_glossary_df = pd.read_csv('../data/processed/glossary/html_cleaned/Ady-Ady_AP.csv')

rus_kbd_glossary_df = rus_kbd_glossary_df.rename(
    columns={
        '#name': 'word',
        'Урыс-адыгэ школ псалъалъэ (Rus-Ady) (eastern)': 'description'}
)
kbd_rus_glossary_df = kbd_rus_glossary_df.rename(
    columns={
        '#name': 'word',
        'Адыгэбзэ псалъалъэ (Ady) (eastern)': 'description'}
)

kbd_rus_glossary_df['word'] = kbd_rus_glossary_df['word'].apply(lambda x: x.lower().replace('i', 'I'))

import nltk

all_sents = set()

for word, desc in kbd_rus_glossary_df[['word', 'description']].values:
    sents = nltk.sent_tokenize(desc)
    all_sents.update({f'{word} - {s}' for s in sents})

for word, desc in rus_kbd_glossary_df[['word', 'description']].values:
    sents = nltk.sent_tokenize(desc)
    all_sents.update({f'{word} - {s}' for s in sents})

all_sents

with open('../data/processed/sents_from_glossary.txt', 'w') as f:
    f.write('\n'.join(sorted(all_sents)))

df = pd.concat([rus_kbd_glossary_df, kbd_rus_glossary_df])

df['desc_rows'] = df['description'].apply(lambda x: x.split('\n'))
df['desc_sents'] = df['description'].apply(lambda x: nltk.sent_tokenize(x))

df
