==> notebooks/02_00_text_analyze_bigram_extract.ipynb <==
import os
from collections import Counter

import nltk

file_dir = '../data/processed'
file_name = 'oshhamaho.txt'
# file_name = 'apkbr_ru.txt'
# file_name = 'elgkbr_ru.txt'
file_path = os.path.join(file_dir, file_name)

BUF_SIZE = 100000
cnt = Counter()
with open(file_path) as f:
    tmp_raw = f.read(BUF_SIZE)
    while tmp_raw:
        tokens = nltk.word_tokenize(tmp_raw)
        text = nltk.Text(tokens)
        clc = text.collocation_list(5000)
        cnt.update(clc)
        tmp_raw = f.read(BUF_SIZE)

import pandas as pd

data = [
    {
        'w_bigram': ' '.join(bigram),
        'freq': freq
    }
    for bigram, freq in cnt.items()
    if freq > 2
]

df = pd.DataFrame(data)
df = df.sort_values('freq', ascending=False)

df

# для дальнейшей работы сохраним полученные данные в csv файл
df.to_csv(
    os.path.join('../data/processed/word_bigrams', file_name.replace('.txt', '.csv')),
    index=False,
)

from nltk.probability import ConditionalFreqDist, ConditionalProbDist, MLEProbDist

cfdist = ConditionalFreqDist()

for bigram, freq in cnt.items():
    word1, word2 = bigram
    cfdist[word1][word2] = freq

cpdist = ConditionalProbDist(cfdist, MLEProbDist)


def get_prob(word1):
    prob_dict = {}
    for word_2 in cpdist[word1].samples():
        prob_dict[word_2] = cpdist[word1].prob(word_2) * 100

    return prob_dict


sorted(get_prob('псалъэ').items(), key=lambda x: x[1], reverse=True)

```python
[('гуапэ', 14.814814814814813),
 ('дахэ', 9.25925925925926),
 ('быдэ', 6.481481481481481),
 ('шэрыуэхэмрэ', 5.555555555555555),
 ('куэд', 4.62962962962963),
 ('къызэринэкIащ', 2.7777777777777777),
 ...
]
```
Мы здесь видим, что самые вероятные слова после `псалъэ` это `гуапэ`, `дахэ`, `быдэ` и т.д.

как мы можем использовать полученные данные?
- для определения того, какие слова чаще всего встречаются вместе
- для предсказания следующего слова с учетом предыдущего
чем это может быть полезно?
- автодополнение при вводе текста
- для автоматического исправления опечаток


Подобное можно еще сделать по токенам, а не по словам.
Тогда мы можем получить вариации слов с учетом падежей, числа и т.д.
То есть зная начало слова, мы можем предсказать наиболее вероятные варианты слова.
Что позволит более точно делать автодополнение и исправление опечаток.
