==> notebooks/02_01_text_analyze_word_frequent_compare_by_source.ipynb <==
import os
import pandas as pd

input_dir = '../data/processed/word_freqs'

df_freq_apkbr_ru = pd.read_csv(os.path.join(input_dir, 'freq_1000000_apkbr_ru.csv'))
df_freq_elgkbr_ru = pd.read_csv(os.path.join(input_dir, 'freq_1000000_elgkbr_ru.csv'))
df_freq_oshhamaho = pd.read_csv(os.path.join(input_dir, 'freq_1000000_oshhamaho.csv'))


top_count = 100
df_freq_apkbr_ru_top = df_freq_apkbr_ru.sort_values(by='freq', ascending=False).head(top_count)
df_freq_elgkbr_ru_top = df_freq_elgkbr_ru.sort_values(by='freq', ascending=False).head(top_count)
df_freq_oshhamaho_top = df_freq_oshhamaho.sort_values(by='freq', ascending=False).head(top_count)

# посмотрим на слова, которые есть в apkbr_ru, но нет в elgkbr_ru и oshhamaho
apkbr_ru = df_freq_apkbr_ru_top.merge(
    df_freq_elgkbr_ru_top, 
    how='left', 
    left_on='word', 
    right_on='word', 
    suffixes=('_apkbr_ru', '_elgkbr_ru')
).merge(
    df_freq_oshhamaho_top, 
    how='left', 
    left_on='word', 
    right_on='word', 
)
apkbr_ru.rename(columns={'freq': 'freq_oshhamaho'}, inplace=True)
apkbr_ru_uniq = apkbr_ru[
    (apkbr_ru.freq_elgkbr_ru.isna()) & 
    (apkbr_ru.freq_oshhamaho.isna())
]
apkbr_ru_uniq

Топ уникальных слов из apkbr_ru (по сравнению с топами elgkbr_ru и oshhamaho) следующий:
"къалъхуащ"
"щыта"
"дунейпсо"
"щIыхь"
"ящыщ"

elgkbr_ru = df_freq_elgkbr_ru_top.merge(
    df_freq_apkbr_ru_top, 
    how='left', 
    left_on='word', 
    right_on='word', 
    suffixes=('_elgkbr_ru', '_apkbr_ru')
).merge(
    df_freq_oshhamaho_top, 
    how='left', 
    left_on='word', 
    right_on='word', 
)
elgkbr_ru.rename(columns={'freq': 'freq_oshhamaho'}, inplace=True)
elgkbr_ru_uniq = elgkbr_ru[
    (elgkbr_ru.freq_apkbr_ru.isna()) & 
    (elgkbr_ru.freq_oshhamaho.isna())
]
elgkbr_ru_uniq

Топ уникальных слов из elgkbr_ru (по сравнению с топами apkbr_ru и oshhamaho) следующий:
"унафэщI"
"КIуэкIуэ"
"Кавказ"
"щIыналъэ"
"лэжьыгъэм"
"хэтащ"
"IэнатIэм"
"ящыщщ"
"сом"
"Федерацэм"
"Апхуэдэу"
"къыхигъэщащ"
"къуажэм"

oshhamaho = df_freq_oshhamaho_top.merge(
    df_freq_apkbr_ru_top, 
    how='left', 
    left_on='word', 
    right_on='word', 
    suffixes=('_oshhamaho', '_apkbr_ru')
).merge(
    df_freq_elgkbr_ru_top, 
    how='left', 
    left_on='word', 
    right_on='word', 
)
oshhamaho.rename(columns={'freq': 'freq_elgkbr_ru'}, inplace=True)
oshhamaho_uniq = oshhamaho[
    (oshhamaho.freq_apkbr_ru.isna()) & 
    (oshhamaho.freq_elgkbr_ru.isna())
]
oshhamaho_uniq

Когда топ уникальных слов из oshhamaho (по сравнению с топами apkbr_ru и elgkbr_ru) более сбалансирован, и правдивее отражает реальную популярность слов без перекосов:

"уэ"
"Сыт"
"Уи"
".."
"*"
"щхьэр"
"Уэ"
"щIалэ"
"щхьэ"
"зыри"
"дэ"
"гум"
"гур"
"гуэр"
"иджы"
"гу"
"хэт"
"гъащIэм"
"иджыри"
"нэхърэ"
"псалъэ"
"цIыкIу"
"игу"
"фIэкIа"
"Зы"
"гуэрым"
"псэр"
"фIэщ"


Видим что популярность слов из новостных источников идут с тематическими перекосами.

То есть, популярные слова из источников в целом пересекаются, но если посмотреть уникальные слова из топа которых нет в топах из соседних источников - то можем увидеть "выбросы" для кажого источника. 

Поэтому для анализа "среднего" текста, нам больше подойдут результаты из источника oshhamaho, так как там нет таких явных перекосов в тематике.
