==> notebooks/02_03_tokenizer_wordpiece.ipynb <==
from tokenizers import Tokenizer
from tokenizers.models import WordPiece
from tokenizers.trainers import WordPieceTrainer
from tokenizers.pre_tokenizers import Whitespace

tokenizer = Tokenizer(WordPiece(unk_token="[UNK]"))

tokenizer.pre_tokenizer = Whitespace()

trainer = WordPieceTrainer(
    vocab_size=2000,
    min_frequency=2,
    special_tokens=["[PAD]", "[UNK]", "[CLS]", "[SEP]", "[MASK]"]
)

files = [ 
    '../data/processed/oshhamaho.txt',
]

tokenizer.train(files, trainer)

sorted(tokenizer.get_vocab().items(), key=lambda x: x[1])

import pandas as pd
words_df = pd.read_csv('../data/processed/word_freqs/freq_1000000_oshhamaho.csv')
words_df['tokens'] = words_df['word'].apply(lambda x: '-'.join([t.replace('##', '') for t in tokenizer.encode(x).tokens]))
words_df
