==> notebooks/02_06_text_analyze_word_groups.ipynb <==
import os
import pandas as pd
import csv

token_dist_dir = '../data/processed/token_distribution'
word_groups_dir = '../data/processed/word_groups'
os.makedirs(word_groups_dir, exist_ok=True)

stem_df = pd.read_csv(f'{token_dist_dir}/stem_candidates.csv.gz', sep=',', quoting=csv.QUOTE_MINIMAL, header=0, compression='gzip')

import json
word_groups_by_template = dict()
for group_name, group_df in stem_df.groupby('template'):
    if len(group_df) > 2:
        word_groups_by_template[group_name] = sorted(set(group_df['word'].to_list()))

with open(f'{word_groups_dir}/word_groups_by_template.json', 'w') as f:
    json.dump(word_groups_by_template, f, ensure_ascii=False, indent=4) 

import re

stem_df['short_template'] = stem_df['template'].apply(lambda x: re.sub(r'\?{1,}', '*', x))

word_groups_by_short_template = dict()
for group_name, group_df in stem_df.groupby('short_template'):
    if group_name == '*':
        continue

    if len(group_df) > 10:
        word_groups_by_short_template[group_name] = sorted(set(group_df['word'].to_list()))

with open(f'{word_groups_dir}/word_groups_by_short_template.json', 'w') as f:
    json.dump(word_groups_by_short_template, f, ensure_ascii=False, indent=4)

word_groups_by_stem = dict()

for group_name, group_df in stem_df.groupby('stem'):
    if len(group_df) > 10:
        word_groups_by_stem[group_name] = sorted(set(group_df['word'].to_list()))

with open(f'{word_groups_dir}/word_groups_by_stem.json', 'w') as f:
    json.dump(word_groups_by_stem, f, ensure_ascii=False, indent=4)
