==> notebooks/02_08_text_analyze_word_dawg.ipynb <==
После визуализации процесса токенизации слов с помощью конечного автомата, мы переходим к использованию префиксных деревьев (trie), что позволяет нам ещё более глубоко анализировать и оптимизировать поиск и обработку слов в больших текстовых корпусах. 

Префиксное дерево ускоряет поиск слов по начальным буквам или фрагментам, обеспечивая эффективную и точную обработку языковых данных, что особенно важно при работе с агглютинативными языками и комплексными морфологическими структурами.

with open('../data/processed/word_freqs/freq_1000000_oshhamaho.txt') as f:
    words = f.read().split('\n')

import ahocorasick
trie = ahocorasick.Automaton(key_type="str", value_type="int")

for id, word in enumerate(sorted(words)):
    trie.add_word(word, id)

trie.make_automaton()

list(trie.keys('къыгурыIуэртэкъым', '?', ahocorasick.MATCH_AT_MOST_PREFIX))

def stem_word(word, trie):
    prefixes = list(trie.keys(word, '?', ahocorasick.MATCH_AT_MOST_PREFIX))
    if not prefixes:
        return word
    
    print(prefixes)
    avg_length = sum(len(p) for p in prefixes) / len(prefixes)
    
    stems = [p for p in prefixes if len(p) > avg_length][:-1]
    return stems

print(stem_word('къыщIыщIэтIысхьар', trie))

list(trie.keys('къы???тэкъым', '?', ahocorasick.MATCH_AT_LEAST_PREFIX))

trie.longest_prefix('къыгурыIуэртэкъыми')

from tqdm import tqdm
from collections import Counter

cnt = Counter()

for word in tqdm(sorted(words)):
    cnt.update(list(trie.keys(word, '?', ahocorasick.MATCH_AT_MOST_PREFIX))[3:])

most_common_words = sorted([(word, freq)  for word, freq in cnt.items() if 5 < freq < 200], key=lambda x: x[1], reverse=True)

most_common_words

len(most_common_words)
