==> notebooks/02_11_spacy_vocab.ipynb <==
import pandas as pd

df_oshhamaho = pd.read_csv('../data/processed/word_freqs/freq_1000000_oshhamaho.csv')
df_elgkbr_ru = pd.read_csv('../data/processed/word_freqs/freq_1000000_elgkbr_ru.csv')
df_apkbr_ru = pd.read_csv('../data/processed/word_freqs/freq_1000000_apkbr_ru.csv')

oshhamaho_words = set(df_oshhamaho['word'].tolist())
elgkbr_ru_words = set(df_elgkbr_ru['word'].tolist())
apkbr_ru_words = set(df_apkbr_ru['word'].tolist())

from spacy.vocab import Vocab

vocab = Vocab(strings=oshhamaho_words.union(elgkbr_ru_words).union(apkbr_ru_words))

vocab.to_disk('../data/processed/spacy/vocab')

import fasttext

kbd_model = fasttext.load_model('../data/processed/embeddings/fasttext_skipgram_kbd_100.bin')

from tqdm import tqdm

for word in tqdm(vocab.strings):
    vocab.set_vector(word, kbd_model.get_word_vector(word))

import gzip

with gzip.open('../data/processed/spacy/kbd_with_vectors.bin.gz', 'wb') as f:
    f.write(vocab.to_bytes())
