==> notebooks/04_collect_word_for_morph.ipynb <==
import fasttext

model = fasttext.load_model('../data/processed/embeddings/fasttext_skipgram_kbd_100.bin')

import pandas as pd

df = pd.read_csv('../data/external/morph/kbd_verb_глагол.csv')

df['key'] = df['key'].str.replace('i', 'I')
df

from tqdm import tqdm

for q in tqdm(df['key']):
    with open(f'../data/external/morph/{q}.txt', 'w') as f:
        f.write('\n'.join([word for score, word in model.get_nearest_neighbors(q, k=50)]))


q = 'джэн'
with open(f'../data/external/morph/{q}.txt', 'w') as f:
    f.write('\n'.join([word for score, word in model.get_nearest_neighbors(q, k=100)]))
