==> notebooks/04_01_ embedding_alignment.ipynb <==
import fasttext

kbd_model = fasttext.load_model('../data/processed/embeddings/fasttext_skipgram_kbd_300.bin')

import gensim.downloader as api
import numpy as np
from gensim.models import KeyedVectors
from scipy.linalg import orthogonal_procrustes

ru_model: KeyedVectors = api.load('word2vec-ruscorpora-300')  # Русские эмбеддинги
flat_keys = {key.split('_')[0]: key for key in ru_model.index_to_key}
# ru_model.get_vector(flat_keys['великан'])

import pandas as pd

rus_kbd_glossary_df = pd.read_csv('../data/processed/glossary/html_cleaned/Rus-Ady_UASP.csv')
kbd_rus_glossary_df = pd.read_csv('../data/processed/glossary/html_cleaned/Ady-Ady_AP.csv')

rus_kbd_glossary_df = rus_kbd_glossary_df.rename(
    columns={
        '#name': 'word',
        'Урыс-адыгэ школ псалъалъэ (Rus-Ady) (eastern)': 'description'}
)
kbd_rus_glossary_df = kbd_rus_glossary_df.rename(
    columns={
        '#name': 'word',
        'Адыгэбзэ псалъалъэ (Ady) (eastern)': 'description'}
)

import nltk

glossary_rus_words = set(rus_kbd_glossary_df['word'].values)
glossary_kbd_words = set(kbd_rus_glossary_df['word'].values)
glossary_kbd_words = {word.lower().replace('i', 'I') for word in glossary_kbd_words}

from tqdm import tqdm

kbd_kv = KeyedVectors(vector_size=300, count=len(glossary_kbd_words))
for word in tqdm(glossary_kbd_words):
    kbd_kv.add_vector(word, kbd_model.get_word_vector(word))


# Поиск соответствий в слов в словаре

Ищем соответствия слов русского языка для слов на кабардинском языке.

kbd_rus_glossary_df['word'] = kbd_rus_glossary_df['word'].apply(lambda x: x.lower().replace('i', 'I'))
kbd_rus_glossary_df['description_flat'] = kbd_rus_glossary_df['description'].apply(
    lambda x: [line for line in x.split('\n') if line][0]
)
kbd_rus_glossary_df['intersection'] = kbd_rus_glossary_df['description_flat'].apply(
    lambda x: set(nltk.word_tokenize(x.lower().replace('i', 'I'))) & glossary_rus_words
)
kbd_rus_glossary_df['intersection_count'] = kbd_rus_glossary_df['intersection'].apply(len)
kbd_rus_glossary_df = kbd_rus_glossary_df[kbd_rus_glossary_df['intersection_count'] == 1]
kbd_rus_glossary_df.head(n=1000)

# маппинг слов: английское слово -> русское слово
bilingual_dict = [
    (kbd_word, list(rus_words)[0])
    for kbd_word, rus_words in kbd_rus_glossary_df[['word', 'intersection']].values
    if list(rus_words)[0] in flat_keys
]

# Извлекаем соответствующие эмбеддинги для каждого слова из пары
X = np.array([kbd_model.get_word_vector(kbd_word) for kbd_word, _ in bilingual_dict])
Y = np.array([ru_model[flat_keys[ru_word]] for _, ru_word in bilingual_dict])

# Вычисляем ортогональное преобразование
R, _ = orthogonal_procrustes(X, Y)

# Применяем преобразование
X_aligned = np.dot(X, R)  # в X_aligned векторы будут похожи на Y
Y_aligned = np.dot(Y, R.T)  # в Y_aligned векторы будут похожи на X

# Собираем выравненные векторы

Чтобы искать соответствия слов на противоположных языках, создадим новые мапы слово -> вектор для каждого языка.

# ключ - слово на русском, значение - выравненный вектор на русском
rus_aligned = KeyedVectors(vector_size=300, count=len(bilingual_dict))
for word, vector in zip([ru_word for _, ru_word in bilingual_dict], Y_aligned):
    rus_aligned.add_vector(word, vector)

# ключ - слово на кабардинском, значение - выравненный вектор на русском
kbd_aligned = KeyedVectors(vector_size=300, count=len(bilingual_dict))
for word, vector in zip([kbd_word for kbd_word, _ in bilingual_dict], X_aligned):
    kbd_aligned.add_vector(word, vector)

from pprint import pprint

pprint(ru_model.similar_by_vector(kbd_aligned['унэ'], topn=5))
pprint(kbd_kv.similar_by_vector(rus_aligned['дом'], topn=5))

pprint(ru_model.similar_by_vector(kbd_aligned['куэбжэ'], topn=5))
pprint(kbd_kv.similar_by_vector(rus_aligned['ворота'], topn=5))
