==> notebooks/04_01_ embedding_glossary_restricted.ipynb <==
import fasttext

vector_size = 100

kbd_model = fasttext.load_model(f'../data/processed/embeddings/fasttext_skipgram_kbd_{vector_size}.bin')

import pandas as pd

kbd_rus_glossary_df = pd.read_csv('../data/processed/glossary/html_cleaned/Ady-Ady_AP.csv')

kbd_rus_glossary_df = kbd_rus_glossary_df.rename(
    columns={
        '#name': 'word',
        'Адыгэбзэ псалъалъэ (Ady) (eastern)': 'description'}
)

glossary_kbd_words = set(kbd_rus_glossary_df['word'].values)
glossary_kbd_words = {word.lower().replace('i', 'I') for word in glossary_kbd_words}

from gensim.models import KeyedVectors
from tqdm import tqdm

# Сохраним вектора только для слов из словаря. (так мы получим похожие слова в словарной форме)
kbd_restricted_glossary_kv = KeyedVectors(vector_size=vector_size, count=len(glossary_kbd_words))
for word in tqdm(glossary_kbd_words):
    kbd_restricted_glossary_kv.add_vector(word, kbd_model.get_word_vector(word))

kbd_restricted_glossary_kv.save(f'../data/processed/embeddings/fasttext_skipgram_kbd_{vector_size}_restricted_glossary.kv')

import json

# Сгруппируем похожие слова для каждого слова из словаря.
glossary_word_groups_by_glossary_words = {}

for word in tqdm(glossary_kbd_words):
    similar_glossary_words = [w for w, _ in kbd_restricted_glossary_kv.most_similar(word, topn=10)]
    glossary_word_groups_by_glossary_words[word] = similar_glossary_words

# "теплъафIэ": [
#     "теплъэншэ",
#     "теплъэ",
#     "теплъаджэ",
#     "фэтеплъэ",
#     "теплъэн",
#     "теплъызэн",
#     "фафIэ",
#     "лыфIэ",
#     "фIэрафIэ",
#     "теплъэгъуей"
# ],
with open('../data/processed/word_groups/glossary_word_groups_by_glossary_words.json', 'w') as f:
    json.dump(glossary_word_groups_by_glossary_words, f, ensure_ascii=False, indent=4)

# соберем слова и векторы из текста отфильтровав слова

file_name = 'freq_1000000_oshhamaho.csv'

df = pd.read_csv(f'../data/processed/word_freqs/{file_name}', sep=',')
words = df[df['freq'] > 1]['word'].values.tolist()

kbd_restricted_kv = KeyedVectors(vector_size=100, count=len(words))
for word in tqdm(words):
    kbd_restricted_kv.add_vector(word, kbd_model.get_word_vector(word))

kbd_restricted_kv.save(f'../data/processed/embeddings/fasttext_skipgram_kbd_{vector_size}_restricted_{file_name}.kv')

# cгруппируем похожие слова для каждого слова из текста 
# (так мы получим похожие слова и различные формы слов)

word_groups_by_glossary_words = {}

for word in tqdm(glossary_kbd_words):
    word_vector = kbd_model.get_word_vector(word)
    similar_words = [w for w, _ in kbd_restricted_kv.similar_by_vector(word_vector, topn=30)]
    word_groups_by_glossary_words[word] = similar_words

# "теплъафIэ": [
#     "теплъафIэ",
#     "теплъафIэр",
#     "теплъафIэт",
#     "теплъафIэм",
#     "теплъа",
#     "теплъэф",
#     "теплъи",
#     "теплъаджэт",
#     "теплъэ…",
#     "теплъэт"
#     ],

with open('../data/processed/word_groups/word_groups_by_glossary_words.json', 'w') as f:
    json.dump(word_groups_by_glossary_words, f, ensure_ascii=False, indent=4)

Если обучить модель на тексте где слова будут заменены на словарную форму или если убрать множество суффиксов, то мы получим более точную и полезную модель.

Можно попробовать смешать подходы и собрать текст с разными модификациями слов:

1. Текст где слова будут без суффиксов.
2. Текст где слова будут без префиксов.
3. Текст где слова будут без суффиксов и префиксов.
4. Текст где слова будут в словарной форме.

То есть, из одного предложения мы можем собрать множество предложений с разными модификациями отделенных слов.

Предположительно, так вес корневой морфемы будет больше и модель будет лучше.
