==> notebooks/04_01_ embedding_word_canonical.ipynb <==
import os

import fasttext

embeddings_dir = '../data/processed/embeddings/'
kbd_model = fasttext.load_model(os.path.join(embeddings_dir, 'fasttext_skipgram_kbd_100.bin'))

from gensim.models import KeyedVectors

kbd_restricted_glossary_kv = KeyedVectors.load(
    os.path.join(embeddings_dir, 'fasttext_skipgram_kbd_100_restricted_glossary.kv'))

from tqdm import tqdm
import pandas as pd

token_dist_dir = '../data/processed/token_distribution/'
stem_df = pd.read_csv(os.path.join(token_dist_dir, 'stem_candidates.csv.gz'), compression='gzip')
stem_df.head(n=100)

word_freq = pd.read_csv('../data/processed/word_freqs/freq_1000000_oshhamaho.csv')
word_freq.head(n=100)

df = pd.merge(stem_df, word_freq, on='word', how='inner')
df = df[df['freq'] > 5]
df.head(n=100)

import ahocorasick

words = df['word'].tolist()

trie = ahocorasick.Automaton(key_type="str", value_type="int")
for id, word in enumerate(sorted(words)):
    trie.add_word(word, id)

trie.make_automaton()

def slice_word_suffix(word, suffix_n=1):
    most_prefix_match = list(trie.keys(word, '?', ahocorasick.MATCH_AT_MOST_PREFIX))
    try:
        return most_prefix_match[-suffix_n - 1]
    except IndexError:
        return None


print(slice_word_suffix('цIыхубзхэри'))

stem_list = set(df['stem'].tolist())


def choose_canonical_by_vector(word):
    word_vector = kbd_model.get_word_vector(word)
    canonical, similarity = kbd_restricted_glossary_kv.similar_by_vector(word_vector, topn=1)[0]
    return canonical, similarity


stem_canonicals = []
for stem in tqdm(stem_list):
    stem_canonicals.append((stem, choose_canonical_by_vector(stem)))

stem_canonicals = dict(stem_canonicals)

from collections import defaultdict
import nltk

trigram_index = defaultdict(list)

restricted_glossary = set(kbd_restricted_glossary_kv.index_to_key)
for glossary_word in tqdm(restricted_glossary):
    for trigram in nltk.ngrams(glossary_word, 3):
        trigram_index[''.join(trigram)].append(glossary_word)

from fuzzywuzzy import fuzz


def get_candidates_by_trigram(word):
    candidates = set()
    for trigram in nltk.ngrams(word, 3):
        candidates.update(trigram_index[''.join(trigram)])

    # print(len(candidates))
    return candidates


def choose_canonical_by_distance(word):
    if word is None:
        return None, 0

    canonical = None
    max_dist = 0
    for candidate in get_candidates_by_trigram(word):
        dist = fuzz.ratio(word, candidate)
        if dist > max_dist:
            max_dist = dist
            canonical = candidate

    score = max_dist / 100
    return canonical, score


print(choose_canonical_by_distance('цIыхубзхэр'))

stem_canonicals_by_distance = {}

for stem in tqdm(stem_list):
    canonical, score = choose_canonical_by_distance(stem)
    stem_canonicals_by_distance[stem] = (canonical, score)

canonicals = []

word_stem = df.set_index('word')['stem'].to_dict()

for word, stem in tqdm(word_stem.items()):
    canonical_by_similar_stem, by_similar_stem_score = stem_canonicals[stem]
    canonical_by_similar_word, by_similar_word_score = choose_canonical_by_vector(word)
    canonical_by_stem_distance, by_stem_distance_score = stem_canonicals_by_distance[stem]
    canonical_by_word_distance, by_word_distance_score = choose_canonical_by_distance(word)

    slice_suffix_1 = slice_word_suffix(word, 1)
    canonical_by_slice_suffix_1, by_slice_suffix_1_score = choose_canonical_by_distance(slice_suffix_1)

    slice_suffix_2 = slice_word_suffix(word, 2)
    canonical_by_slice_suffix_2, by_slice_suffix_2_score = choose_canonical_by_distance(slice_suffix_2)

    slice_suffix_3 = slice_word_suffix(word, 3)
    canonical_by_slice_suffix_3, by_slice_suffix_3_score = choose_canonical_by_distance(slice_suffix_3)

    canonicals.append({
        'word': word,
        'canonical_by_similar_word': canonical_by_similar_word,
        'canonical_by_similar_word_score': by_similar_word_score,
        'canonical_by_word_distance': canonical_by_word_distance,
        'canonical_by_word_distance_score': by_word_distance_score,

        'stem': stem,
        'canonical_by_similar_stem': canonical_by_similar_stem,
        'canonical_by_similar_stem_score': by_similar_stem_score,
        'canonical_by_stem_distance': canonical_by_stem_distance,
        'canonical_by_stem_distance_score': by_stem_distance_score,

        'slice_suffix_1': slice_suffix_1,
        'canonical_by_slice_suffix_1': canonical_by_slice_suffix_1,
        'by_slice_suffix_1_score': by_slice_suffix_1_score,

        'slice_suffix_2': slice_suffix_2,
        'canonical_by_slice_suffix_2': canonical_by_slice_suffix_2,
        'by_slice_suffix_2_score': by_slice_suffix_2_score,

        'slice_suffix_3': slice_suffix_3,
        'canonical_by_slice_suffix_3': canonical_by_slice_suffix_3,
        'by_slice_suffix_3_score': by_slice_suffix_3_score,
    })

canonicals_df = pd.DataFrame(canonicals)

canonicals_df

similarity_columns = canonicals_df.filter(like='score').columns.tolist()
similarity_columns

score_columns = [
    'canonical_by_similar_word_score',
    'canonical_by_word_distance_score',
    'canonical_by_similar_stem_score',
    'canonical_by_stem_distance_score',
    'by_slice_suffix_1_score',
    'by_slice_suffix_2_score',
    # 'by_slice_suffix_3_score'
]

filtered_df = canonicals_df[canonicals_df[score_columns].gt(0.7).all(axis=1)]
filtered_df

canonicals_df[(canonicals_df['canonical_by_similar_word'] == canonicals_df['canonical_by_word_distance'])]

canonicals_df.sort_values(by='canonical_by_slice_suffix_3', ascending=False)

from collections import Counter


text_columns = [
    'slice_suffix_1',
    'slice_suffix_2',
    'slice_suffix_3',

    'canonical_by_similar_word',
    'canonical_by_word_distance',
    'canonical_by_similar_stem',
    'canonical_by_stem_distance',
    'canonical_by_slice_suffix_1',
    'canonical_by_slice_suffix_2',
    'canonical_by_slice_suffix_3',
]

def most_common(row):
    most_common_value = max(Counter(row).items(), key=lambda x: x[1] if x[0] is not None else 0)
    return most_common_value

canonicals_df['most_common'] = canonicals_df[text_columns].apply(most_common, axis=1)
canonicals_df['most_common_value'] = canonicals_df['most_common'].apply(lambda x: x[0])
canonicals_df['most_common_count'] = canonicals_df['most_common'].apply(lambda x: x[1])
canonicals_df.sort_values(by=['most_common_value', 'most_common_count'], ascending=False)
