==> notebooks/04_01_embeddings_train_fasttext.ipynb <==
import os

import nltk

file_dir = '../data/processed'

file_list = [
    'all_book.txt',
    'apkbr_ru.txt',
    'elgkbr_ru.txt',
    'oshhamaho.txt',
]

def clean_sentence(sentence):
    return ' '.join([word for word in nltk.word_tokenize(sentence) if len(word) > 2])


sentences = []
for file in file_list:
    with open(os.path.join(file_dir, file), 'r') as f:
        text = f.read()

    for sent in nltk.sent_tokenize(text):
        if len(sent) > 10:
            sentences.append(clean_sentence(sent))

with open(os.path.join(file_dir, 'all_sentences.txt'), 'w') as f:
    f.write('\n'.join(sentences))

import os
import fasttext

# train_file = os.path.join(file_dir, 'all_sentences.txt')

model = fasttext.train_unsupervised('../data/processed/sentences_by_lang/kbd.txt', model='skipgram', dim=100, minCount=2, epoch=5)

model.get_nearest_neighbors('къыщыщI', k=20)

model_dir = '../data/processed/embeddings'
os.makedirs(model_dir, exist_ok=True)

model.save_model(os.path.join(model_dir, 'fasttext_skipgram_kbd_100.bin'))

model.get_nearest_neighbors('щытехьам', k=100)
