==> notebooks/04_02_gensim_word2vec.ipynb <==
import nltk
import pandas as pd

import ahocorasick

word_freq_df = pd.read_csv('../data/processed/word_freqs/freq_1000000_oshhamaho.csv')
word_freq_df = word_freq_df[word_freq_df['freq'] > 1]
words = set(word_freq_df['word'].values)

trie = ahocorasick.Automaton(key_type="str", value_type="int")
for id, word in enumerate(sorted(words)):
    trie.add_word(word, id)

trie.make_automaton()


def slice_word_suffix(word, min_len=3):
    most_prefix_match = list(trie.keys(word, '?', ahocorasick.MATCH_AT_MOST_PREFIX))
    return [word for word in most_prefix_match if len(word) >= min_len]


import math
from tqdm import tqdm


class Replacer:

    def __init__(self, words=words):
        self.word_slice_map = {}
        for word in words:
            self.word_slice_map[word] = slice_word_suffix(word)

    def replace_by_slice(self, line):
        line_words = nltk.word_tokenize(line)
        lines = set()

        for word in line_words:
            slice_words = self.word_slice_map.get(word, [])
            for slice_word in slice_words:
                if slice_word != word:
                    new_line = line.replace(word, slice_word)

                    sent_tokens = [w for w in nltk.word_tokenize(new_line) if len(w) > 1]
                    lines.add(tuple(sent_tokens))

        return lines

    def replace(self, line):
        lines = set()

        replaced_by_slice = self.replace_by_slice(line)
        if replaced_by_slice:
            lines.update(replaced_by_slice)

        return sorted(lines)


replacer = Replacer()

replacer.replace('Шэч хэмылъу , псэ мызагъэу , цIыхухэм яхуэпэжу , пщIэшхуэ къэзылэжьу дунейм тета')

class MyCorpus:

    def __init__(self, corpus_path, max_iter=math.inf):
        self.corpus_path = corpus_path
        self.iter_count = 0
        self.max_iter_count = max_iter
        with open(self.corpus_path) as f:
            text = f.read()
        self._sentences = sorted(set(nltk.sent_tokenize(text)))

        self.pbar = tqdm(self._sentences)
        self.last_flushed = None

    def __iter__(self):
        for line_1 in self.pbar:
            if self.iter_count > self.max_iter_count:
                break

            lines = replacer.replace(line_1)
            for line in lines:
                self.iter_count += 1
                yield line


corpus_path = '../data/processed/all_sentences.txt'
corpus = MyCorpus(corpus_path)
sentences = list(corpus)


from gensim.models.callbacks import CallbackAny2Vec


class MonitorCallback(CallbackAny2Vec):
    def __init__(self, test_words):
        self._test_words = test_words

    def on_epoch_end(self, model):
        print("Model loss:", model.get_latest_training_loss())  # print loss
        for _word in self._test_words:  # show wv logic changes
            print(_word, model.wv.most_similar(_word))

import gensim.models

monitor = MonitorCallback([
    'бажэ',
    'унэ',
    'бадзэ',
    'уафэ',
    'къуэш',
    'еджэн',
    'шхын',
])

model = gensim.models.FastText(
    vector_size=300,
    workers=16,
    window=7,
    negative=5,
    sg=1
)

model.build_vocab(corpus_iterable=sentences)

model.train(
    corpus_iterable=sentences,
    total_examples=len(sentences),
    epochs=5,
    callbacks=[monitor],
    compute_loss=True
)

model.wv.similar_by_word('гупсысащ')

import os

os.makedirs('../data/processed/embeddings', exist_ok=True)
model.save('../data/processed/embeddings/gensim_word2vec_1000000_oshhamaho.model')
