==> notebooks/04_03_prepare_visualize.ipynb <==
import fasttext
import os

embedding_dir = '../data/processed/embeddings/'
kbd_model = fasttext.load_model(os.path.join(embedding_dir, 'fasttext_skipgram_kbd_300.bin'))

from tqdm import tqdm
import pandas as pd

file_name = 'freq_1000000_oshhamaho.csv'

df = pd.read_csv(f'../data/processed/word_freqs/{file_name}', sep=',')
words = df[df['freq'] > 10]['word'].values.tolist()

vectors = [kbd_model.get_word_vector(word) for word in tqdm(words)]

visual_dir = '../data/processed/embeddings/visualizations/'

vectors_df = pd.DataFrame(vectors)
vectors_df.to_csv(os.path.join(visual_dir, "word_vectors.tsv"), sep='\t', header=False, index=False)

words_df = pd.DataFrame(words)
words_df.to_csv(os.path.join(visual_dir, "word_labels.tsv"), sep='\t', header=False, index=False)

s1 = 'Си гуапэщ я ныбжьэгъугъэм сызэрыригушхуэр абыхэм ящIэну.'
s2 = 'Сэ зэныбжьэгъугъэм абыхэм гуапагъ хэлъщ сызэрыригушхуэр.'

w1 = kbd_model.get_sentence_vector(s1)
w2 = kbd_model.get_sentence_vector(s2)

import numpy as np

distance = np.linalg.norm(w1 - w2)
distance
