==> notebooks/04_08_concordance.ipynb <==
import nltk

with open('../data/processed/oshhamaho.txt', 'r') as f:
    oshhamaho = f.read()
    

short_sents = [sent for sent in nltk.sent_tokenize(oshhamaho) if 60 < len(sent) < 100]

short_sent_text = '\n'.join(sorted(set(short_sents)))

from nltk import Text

text = Text(tokens=nltk.word_tokenize(short_sent_text))

text.concordance('щыщ')
