==> notebooks/04_09_sentence_bert_aign.ipynb <==
from sentence_transformers import SentencesDataset, InputExample

import pandas as pd

example_hash = set()
train_examples = []

sent_pairs_df = pd.read_csv('../data/processed/word_freqs/freq_1000000_oshhamaho_translated.csv')
sent_pairs_df.dropna(inplace=True)
sent_pairs_df = sent_pairs_df[sent_pairs_df['kbd'] != sent_pairs_df['rus']]
sent_pairs_df = sent_pairs_df.sample(10000, random_state=42)

for kbd, rus in sent_pairs_df[['kbd', 'rus']].values:
    train_examples.append(
        InputExample(texts=[kbd, rus], label=1.0)
    )

from sentence_transformers import SentenceTransformer, losses
from torch.utils.data import DataLoader

model = SentenceTransformer('./sbert_from_mlm_bert_21')

# Создание и загрузка датасета
train_dataset = SentencesDataset(examples=train_examples, model=model)
train_dataloader = DataLoader(train_dataset, batch_size=8, shuffle=True)

eval_examples = train_examples[:100]

eval_dataset = SentencesDataset(examples=eval_examples, model=model)

from sentence_transformers.evaluation import EmbeddingSimilarityEvaluator

evaluator = EmbeddingSimilarityEvaluator.from_input_examples(
    examples=eval_examples,
    name='sent_pairs_eval'
)

# Настройка процесса обучения
train_loss = losses.CosineSimilarityLoss(model=model)

# Обучение модели
model.fit(
    train_objectives=[(train_dataloader, train_loss)],
    evaluation_steps=100,
    epochs=1
)

model.save('./sbert_aligned')

from scipy.spatial.distance import cosine

model.encode(
    ['привет', 'пока'],
)

for kbd, rus in sent_pairs_df[['kbd', 'rus']].values:
    v1, v2 = model.encode([kbd, rus])
