==> notebooks/05_00_translate.ipynb <==
from transformers import M2M100Tokenizer, M2M100ForConditionalGeneration

model_ru_kbd = M2M100ForConditionalGeneration.from_pretrained("anzorq/m2m100_1.2B_ft_ru-kbd_63K")
tokenizer_ru_kbd = M2M100Tokenizer.from_pretrained("anzorq/m2m100_418M_ft_ru-kbd_50K_T4", src_lang="ru")

tokenizer_ru_kbd.tokenize('НтIэ, сытыт укъыщIысхуейр?')

def translate(text):
    encoded = tokenizer_ru_kbd(text, return_tensors="pt")
    generated_tokens = model_ru_kbd.generate(**encoded)
    translated = tokenizer_ru_kbd.batch_decode(generated_tokens, skip_special_tokens=True)
    return translated[0]

translate('Я попытался её поймать.')

from transformers import M2M100Tokenizer, M2M100ForConditionalGeneration

model_kbd_ru = M2M100ForConditionalGeneration.from_pretrained("anzorq/m2m100_1.2B_ft_kbd-ru_63K")

tokenizer_kbd_ru = M2M100Tokenizer.from_pretrained("anzorq/m2m100_1.2B_ft_kbd-ru_63K", src_lang="ru")
tokenizer_kbd_ru.tokenize('НтIэ, сытыт укъыщIысхуейр?')

def translate2(text):
    encoded = tokenizer_kbd_ru(text, return_tensors="pt")
    generated_tokens = model_kbd_ru.generate(**encoded)
    translated = tokenizer_kbd_ru.batch_decode(generated_tokens, skip_special_tokens=True)
    return translated[0]

translate2("""НтIэ, сытыт укъыщIысхуейр?""")

import pandas as pd

bigram_df = pd.read_csv('../data/processed/word_bigrams/oshhamaho.csv')
bigram_df.head()
bigram_df['translation'] = bigram_df['w_bigram'].apply(translate2)
bigram_df['corrected'] = bigram_df['translation']
bigram_df.to_csv('../data/processed/word_bigrams/oshhamaho_translated.csv', index=False)

import pandas as pd
import os

cluster_path = '../data/processed/sent_clusters'

with open('../data/processed/sent_clusters/seed_113/10.0_10000_1000/cluster_0_10000_99.txt', 'r') as f:
    sentences = f.read().split('\n')

cluster_df = pd.DataFrame({'sentence': sentences})
cluster_df['translation'] = cluster_df['sentence'].apply(translate2)
cluster_df['corrected'] = cluster_df['translation']

cluster_df
