==> notebooks/05_01_datasets.ipynb <==
# !pip install datasets

from datasets import load_dataset

dataset = load_dataset("anzorq/ru-kbd")

import pandas as pd

df = pd.DataFrame(dataset['train']['translation'])

df.sort_values('kbd')

ru_sent = set(df['ru'].to_list())
kbd_sent = set(df['kbd'].to_list())

'слушать внимательно, ничего не пропуская' in ru_sent
